+++ implementierung/src/worthwhile.model/src/edu/kit/iti/formal/pse/worthwhile/model/CompositeValue.java
@@ -1,5 +1,6 @@
 package edu.kit.iti.formal.pse.worthwhile.model;
 
+import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -19,14 +20,14 @@ public class CompositeValue<T extends Value> extends Value {
 	/**
 	 * The sub-values of this composite value.
 	 */
-	private final Map<Integer, T> subValues;
+	private final Map<BigInteger, T> subValues;
 
 	/**
 	 * Returns the sub-values of this composite value.
 	 * 
 	 * @return The sub-values of this composite value.
 	 */
-	public final Map<Integer, T> getSubValues() {
+	public final Map<BigInteger, T> getSubValues() {
 		return this.subValues;
 	}
 
@@ -36,7 +37,7 @@ public class CompositeValue<T extends Value> extends Value {
 	 * @param subValues
 	 *                The sub-values of this value.
 	 */
-	public CompositeValue(final Map<Integer, T> subValues) {
+	public CompositeValue(final Map<BigInteger, T> subValues) {
 		this.subValues = subValues;
 	}
 
@@ -47,9 +48,9 @@ public class CompositeValue<T extends Value> extends Value {
 	 *                The sub-values of this value.
 	 */
 	public CompositeValue(final T[] subValues) {
-		this.subValues = new HashMap<Integer, T>();
+		this.subValues = new HashMap<BigInteger, T>();
 		for (int i = 0; i < subValues.length; i++) {
-			this.subValues.put(i, subValues[i]);
+			this.subValues.put(BigInteger.valueOf(i), subValues[i]);
 		}
 	}
 
@@ -64,11 +65,11 @@ public class CompositeValue<T extends Value> extends Value {
 	 * @return a new instance of the {@link CompositeValue} class with the value at index i replaced by (or newly
 	 *         set to) the specified value.
 	 */
-	public CompositeValue<T> replaceValue(int index, T newValue) {
+	public CompositeValue<T> replaceValue(BigInteger index, T newValue) {
 		// Copy the current values to a new composite value
-		Map<Integer, T> newValues = new HashMap<Integer, T>();
+		Map<BigInteger, T> newValues = new HashMap<BigInteger, T>();
 		
-		for (Integer idx : this.subValues.keySet()) {
+		for (BigInteger idx : this.subValues.keySet()) {
 			newValues.put(idx, this.getSubValues().get(idx));
 		}
 		
@@ -97,7 +98,7 @@ public class CompositeValue<T extends Value> extends Value {
 	public final int hashCode() {
 		int result = 0;
 
-		for (Integer index : this.subValues.keySet()) {
+		for (BigInteger index : this.subValues.keySet()) {
 			result += this.subValues.get(index).hashCode();
 		}
